requirements.txt
matplotlib==2.1.0
numpy==1.13.3
inflect==0.2.5
Unidecode==1.0.22
pytorch==1.9.0
pytorch-lightning=1.4.2
pillow
